Refactored TUI project layout
