--- software/script/requirements.txt
@@ -1,3 +1,0 @@
-pyserial==3.5
-colorama==0.4.6
-prompt-toolkit==3.0.39
